Reworked the orderbook to work correctly.
Still a pretty naive implementation but works much better.  One of the significant
issues with the original implementation was that the Orderbook() didn't know the
number of levels it was managing and incorrect books could occur.

The result is OrderBooks now need the number of levels passed when they are created.
for the CME, this can be derived from the secdef.dat.gz file on their web site.

Added a module, SecDef(), to handle loading/parsing the secdef.dat.gz into a
dictionary indexed by the security id.